Remove the obsolete AndroidClientHandler class and related code
