+++ Documentation/docs-mobile/TOC.yml
@@ -258,12 +258,6 @@
         href: messages/xa1028.md
       - name: XA1029
         href: messages/xa1029.md
-      - name: XA1031
-        href: messages/xa1031.md
-      - name: XA1032
-        href: messages/xa1032.md
-      - name: XA1033
-        href: messages/xa1033.md
       - name: XA1035
         href: messages/xa1035.md
       - name: XA1036
