--- Documentation/docs-mobile/messages/xa1031.md
@@ -1,28 +1,0 @@
----
-title: .NET for Android error XA1031
-description: XA1031 error code
-ms.date: 04/11/2024
-f1_keywords:
-  - "XA1031"
----
-
-# .NET for Android error XA1031
-
-## Example messages
-
-```
-The 'AndroidHttpClientHandlerType' property value 'Foo.Bar.HttpHander, MyApp' must derive from 'System.Net.Http.HttpMessageHandler'.
-Please change the value to an assembly-qualifed type name which inherits from '{1}' or remove the property completely.
-```
-
-```
-The 'AndroidHttpClientHandlerType' property value '.NET for Android.Net.AndroidClientHandler' must not derive from 'System.Net.Htt.HttpClientHandler'.
-Please change the value to an assembly-qualifed type name which inherits from 'System.Net.Http.HttpMessageHandler' or remove the property completely.
-```
-
-## Solution
-
-Edit your csproj directly and change the 'AndroidHttpClientHandlerType' to
-a valid value.
-
-Valid values can be found at `~/android/deploy-test/building-apps/build-properties.md#AndroidHttpClientHandlerType`.
