--- Documentation/docs-mobile/messages/xa1032.md
@@ -1,22 +1,0 @@
----
-title: .NET for Android error XA1032
-description: XA1032 error code
-ms.date: 04/11/2024
-f1_keywords:
-  - "XA1032"
----
-
-# .NET for Android error XA1032
-
-## Example messages
-
-```
-Failed to resolve '{0}' from '{1}'. Please check your `AndroidHttpClientHandlerType` setting.
-```
-
-## Solution
-
-Edit your csproj directly and change the 'AndroidHttpClientHandlerType' to
-a valid value.
-
-Valid values can be found at `~/android/deploy-test/building-apps/build-properties.md#AndroidHttpClientHandlerType`.
