--- Documentation/docs-mobile/messages/xa1033.md
@@ -1,22 +1,0 @@
----
-title: .NET for Android error XA1033
-description: XA1033 error code
-ms.date: 04/11/2024
-f1_keywords:
-  - "XA1033"
----
-
-# .NET for Android error XA1033
-
-## Example messages
-
-```
-Could not resolve '{0}'. Please check your `AndroidHttpClientHandlerType` setting.
-```
-
-## Solution
-
-Edit your csproj directly and change the 'AndroidHttpClientHandlerType' to
-a valid value.
-
-Valid values can be found at `~/android/deploy-test/building-apps/build-properties.md#AndroidHttpClientHandlerType`.
