+++ src/Mono.Android/Android.Runtime/AndroidEnvironment.cs
@@ -26,8 +26,6 @@ namespace Android.Runtime {
 		static IX509TrustManager? sslTrustManager;
 		static KeyStore? certStore;
 		static object lock_ = new object ();
-		[DynamicallyAccessedMembers (DynamicallyAccessedMemberTypes.PublicParameterlessConstructor)]
-		static Type? httpMessageHandlerType;
 
 		static void SetupTrustManager ()
 		{
@@ -253,66 +251,9 @@ namespace Android.Runtime {
 		// This is invoked by
 		// System.Net.Http.dll!System.Net.Http.HttpClient.cctor
 		// DO NOT REMOVE
-		[DynamicDependency (DynamicallyAccessedMemberTypes.PublicParameterlessConstructor, typeof (Xamarin.Android.Net.AndroidMessageHandler))]
 		static HttpMessageHandler GetHttpMessageHandler ()
 		{
-			[UnconditionalSuppressMessage ("Trimming", "IL2057", Justification = "Preserved by the MarkJavaObjects trimmer step.")]
-			[return: DynamicallyAccessedMembers (DynamicallyAccessedMemberTypes.PublicParameterlessConstructor)]
-			static Type? TypeGetType (string typeName) =>
-				Type.GetType (typeName, throwOnError: false);
-
-			if (httpMessageHandlerType is null) {
-				var handlerTypeName = Environment.GetEnvironmentVariable ("XA_HTTP_CLIENT_HANDLER_TYPE")?.Trim ();
-				Type? handlerType = null;
-				if (!String.IsNullOrEmpty (handlerTypeName))
-					handlerType = TypeGetType (handlerTypeName);
-
-				if (handlerType is null || !IsAcceptableHttpMessageHandlerType (handlerType)) {
-					handlerType = GetFallbackHttpMessageHandlerType ();
-				}
-
-				httpMessageHandlerType = handlerType;
-			}
-
-			return (HttpMessageHandler) Activator.CreateInstance (httpMessageHandlerType)
-				?? throw new InvalidOperationException ($"Could not create an instance of HTTP message handler type {httpMessageHandlerType.AssemblyQualifiedName}");
+			return new Xamarin.Android.Net.AndroidMessageHandler ();
 		}
-
-		static bool IsAcceptableHttpMessageHandlerType (Type handlerType)
-		{
-			if (Extends (handlerType, "System.Net.Http.HttpClientHandler, System.Net.Http")) {
-				// It's not possible to construct HttpClientHandler in this method because it would cause infinite recursion
-				// as HttpClientHandler's constructor calls the GetHttpMessageHandler function
-				Logger.Log (LogLevel.Warn, "MonoAndroid", $"The type {handlerType.AssemblyQualifiedName} cannot be used as the native HTTP handler because it is derived from System.Net.Htt.HttpClientHandler. Use a type that extends System.Net.Http.HttpMessageHandler instead.");
-				return false;
-			}
-			if (!Extends (handlerType, "System.Net.Http.HttpMessageHandler, System.Net.Http")) {
-				Logger.Log (LogLevel.Warn, "MonoAndroid", $"The type {handlerType.AssemblyQualifiedName} set as the default HTTP handler is invalid. Use a type that extends System.Net.Http.HttpMessageHandler.");
-				return false;
-			}
-
-			return true;
-		}
-
-		static bool Extends (
-				Type handlerType,
-				[DynamicallyAccessedMembers (DynamicallyAccessedMemberTypes.PublicParameterlessConstructor)]
-				string baseTypeName)
-		{
-			var baseType = Type.GetType (baseTypeName, throwOnError: false);
-			return baseType?.IsAssignableFrom (handlerType) ?? false;
-		}
-
-		[return: DynamicallyAccessedMembers (DynamicallyAccessedMemberTypes.PublicParameterlessConstructor)]
-		static Type GetFallbackHttpMessageHandlerType ()
-		{
-			const string typeName = "Xamarin.Android.Net.AndroidMessageHandler, Mono.Android";
-			var handlerType = Type.GetType (typeName, throwOnError: false)
-				?? throw new InvalidOperationException ($"The {typeName} was not found. The type was probably linked away.");
-
-			Logger.Log (LogLevel.Info, "MonoAndroid", $"Using {typeName} as the native HTTP message handler.");
-			return handlerType;
-		}
-
 	}
 }
