+++ src/Mono.Android/Xamarin.Android.Net/AndroidMessageHandler.cs
@@ -421,9 +421,6 @@ namespace Xamarin.Android.Net
 			return _serverCertificateCustomValidator?.HostnameVerifier;
 		}
 
-		internal IHostnameVerifier? GetSSLHostnameVerifierInternal (HttpsURLConnection connection)
-			=> GetSSLHostnameVerifier (connection);
-
 		/// <summary>
 		/// Creates, configures and processes an asynchronous request to the indicated resource.
 		/// </summary>
@@ -514,9 +511,6 @@ namespace Xamarin.Android.Net
 			}
 		}
 
-		internal Task <HttpResponseMessage> SendAsyncInternal (HttpRequestMessage request, CancellationToken cancellationToken)
-			=> SendAsync (request, cancellationToken);
-
 		protected virtual async Task <Java.Net.Proxy?> GetJavaProxy (Uri destination, CancellationToken cancellationToken)
 		{
 			var proxy = Java.Net.Proxy.NoProxy;
@@ -538,9 +532,6 @@ namespace Xamarin.Android.Net
 
 			return proxy;
 		}
-
-		internal Task <Java.Net.Proxy?> GetJavaProxyInternal (Uri destination, CancellationToken cancellationToken)
-			=> GetJavaProxy (destination, cancellationToken);
 
 		Task <HttpResponseMessage?> ProcessRequest (HttpRequestMessage request, URL javaUrl, HttpURLConnection httpConnection, CancellationToken cancellationToken, RequestRedirectionState redirectState)
 		{
@@ -614,9 +605,6 @@ namespace Xamarin.Android.Net
 					stream.Seek (0, SeekOrigin.Begin);
 			}
 		}
-
-		internal Task WriteRequestContentToOutputInternal (HttpRequestMessage request, HttpURLConnection httpConnection, CancellationToken cancellationToken)
-			=> WriteRequestContentToOutput (request, httpConnection, cancellationToken);
 
 		async Task <HttpResponseMessage?> DoProcessRequest (HttpRequestMessage request, URL javaUrl, HttpURLConnection httpConnection, CancellationToken cancellationToken, RequestRedirectionState redirectState)
 		{
@@ -1050,9 +1038,6 @@ namespace Xamarin.Android.Net
 			return Task.CompletedTask;
 		}
 
-		internal Task SetupRequestInternal (HttpRequestMessage request, HttpURLConnection conn)
-			=> SetupRequest (request, conn);
-
 		/// <summary>
 		/// Configures the key store. The <paramref name="keyStore"/> parameter is set to instance of <see cref="KeyStore"/>
 		/// created using the <see cref="KeyStore.DefaultType"/> type and with populated with certificates provided in the <see cref="TrustedCerts"/>
@@ -1066,9 +1051,6 @@ namespace Xamarin.Android.Net
 
 			return keyStore;
 		}
-
-		internal KeyStore? ConfigureKeyStoreInternal (KeyStore? keyStore)
-			=> ConfigureKeyStore (keyStore);
 
 		/// <summary>
 		/// Create and configure an instance of <see cref="KeyManagerFactory"/>. The <paramref name="keyStore"/> parameter is set to the
@@ -1086,9 +1068,6 @@ namespace Xamarin.Android.Net
 			return null;
 		}
 
-		internal KeyManagerFactory? ConfigureKeyManagerFactoryInternal (KeyStore? keyStore)
-			=> ConfigureKeyManagerFactoryInternal (keyStore);
-
 		/// <summary>
 		/// Create and configure an instance of <see cref="TrustManagerFactory"/>. The <paramref name="keyStore"/> parameter is set to the
 		/// return value of the <see cref="ConfigureKeyStore"/> method, so it might be null if the application overrode the method and provided
@@ -1105,9 +1084,6 @@ namespace Xamarin.Android.Net
 
 			return null;
 		}
-
-		internal TrustManagerFactory? ConfigureTrustManagerFactoryInternal (KeyStore? keyStore)
-			=> ConfigureTrustManagerFactory (keyStore);
 
 		async Task <HttpURLConnection> SetupRequestInternal (HttpRequestMessage request, URLConnection conn)
 		{
@@ -1161,9 +1137,6 @@ namespace Xamarin.Android.Net
 		{
 			return null;
 		}
-
-		internal SSLSocketFactory? ConfigureCustomSSLSocketFactoryInternal (HttpsURLConnection connection)
-			=> ConfigureCustomSSLSocketFactoryInternal (connection);
 
 		void SetupSSL (HttpsURLConnection? httpsConnection, HttpRequestMessage requestMessage)
 		{
