+++ src/Xamarin.Android.Build.Tasks/Microsoft.Android.Sdk/targets/Microsoft.Android.Sdk.NativeAOT.targets
@@ -263,7 +263,6 @@ This file contains the NativeAOT-specific MSBuild logic for .NET for Android.
 
     <GenerateNativeAotEnvironmentAssemblerSources
         Environments="@(AndroidEnvironment);@(LibraryEnvironments)"
-        HttpClientHandlerType="$(AndroidHttpClientHandlerType)"
         OutputSources="@(_PrivateEnvironmentAssemblySource)"
         RID="$(RuntimeIdentifier)" />
