+++ src/Xamarin.Android.Build.Tasks/Microsoft.Android.Sdk/targets/Microsoft.Android.Sdk.TypeMap.Trimmable.targets
@@ -200,7 +200,6 @@
         OutputDirectory="$(IntermediateOutputPath)android"
         TargetName="$(TargetName)"
         Environments="@(_EnvironmentFiles)"
-        HttpClientHandlerType="$(AndroidHttpClientHandlerType)"
         EnableSGenConcurrent="$(AndroidEnableSGenConcurrent)" />
 
     <!-- Copy generated manifest to expected location -->
