+++ src/Xamarin.Android.Build.Tasks/Tasks/GenerateNativeApplicationConfigSources.cs
@@ -85,10 +85,9 @@ namespace Xamarin.Android.Tasks
 
 		AndroidRuntime androidRuntime;
 
-		internal static void AddDefaultEnvironmentVariables (EnvironmentBuilder envBuilder, string? httpClientHandlerType, bool enableSGenConcurrent)
+		internal static void AddDefaultEnvironmentVariables (EnvironmentBuilder envBuilder, bool enableSGenConcurrent)
 		{
 			envBuilder.AddDefaultMonoDebug ();
-			envBuilder.AddHttpClientHandlerType (httpClientHandlerType);
 			envBuilder.AddMonoGcParams (enableSGenConcurrent);
 		}
 
@@ -121,7 +120,7 @@ namespace Xamarin.Android.Tasks
 			}
 
 			if (androidRuntime != Xamarin.Android.Tasks.AndroidRuntime.NativeAOT) {
-				AddDefaultEnvironmentVariables (envBuilder, HttpClientHandlerType, EnableSGenConcurrent);
+				AddDefaultEnvironmentVariables (envBuilder, EnableSGenConcurrent);
 			} else {
 				// NativeAOT sets all the environment variables from Java, we don't want to repeat that
 				// process in the native code. This is just a precaution, because NativeAOT builds should
