+++ src/Xamarin.Android.Build.Tasks/Tests/Xamarin.Android.Build.Tests/Tasks/LinkerTests.cs
@@ -161,79 +161,6 @@ namespace Xamarin.Android.Build.Tests
 			return assm;
 		}
 
-		private void PreserveCustomHttpClientHandler (
-				string handlerType,
-				string handlerAssembly,
-				string testProjectName,
-				string assemblyPath,
-				TrimMode trimMode,
-				AndroidRuntime runtime)
-		{
-			const bool isRelease = true;
-			if (IgnoreUnsupportedConfiguration (runtime, release: isRelease)) {
-				return;
-			}
-			testProjectName += trimMode.ToString ();
-
-			var class_library = new XamarinAndroidLibraryProject {
-				IsRelease = isRelease,
-				ProjectName = "MyClassLibrary",
-				Sources = {
-					new BuildItem.Source ("MyCustomHandler.cs") {
-						TextContent = () => """
-							class MyCustomHandler : System.Net.Http.HttpMessageHandler
-							{
-								protected override Task <HttpResponseMessage> SendAsync (HttpRequestMessage request, CancellationToken cancellationToken) =>
-									throw new NotImplementedException ();
-							}
-						"""
-					},
-					new BuildItem.Source ("Bar.cs") {
-						TextContent = () => "public class Bar { }",
-					}
-				}
-			};
-			class_library.SetRuntime (runtime);
-			using (var libBuilder = CreateDllBuilder ($"{testProjectName}/{class_library.ProjectName}")) {
-				Assert.IsTrue (libBuilder.Build (class_library), $"Build for {class_library.ProjectName} should have succeeded.");
-			}
-
-			var proj = new XamarinAndroidApplicationProject {
-				ProjectName = "MyApp",
-				IsRelease = isRelease,
-				TrimModeRelease = trimMode,
-				Sources = {
-					new BuildItem.Source ("Foo.cs") {
-						TextContent = () => "public class Foo : Bar { }",
-					}
-				}
-			};
-			proj.SetRuntime (runtime);
-			proj.AddReference (class_library);
-			proj.AddReferences ("System.Net.Http");
-			string handlerTypeFullName = string.IsNullOrEmpty(handlerAssembly) ? handlerType : handlerType + ", " + handlerAssembly;
-			proj.SetProperty (proj.ActiveConfigurationProperties, "AndroidHttpClientHandlerType", handlerTypeFullName);
-			proj.MainActivity = proj.DefaultMainActivity.Replace ("base.OnCreate (bundle);", "base.OnCreate (bundle);\nvar client = new System.Net.Http.HttpClient ();");
-			using (var b = CreateApkBuilder ($"{testProjectName}/{proj.ProjectName}")) {
-				Assert.IsTrue (b.Build (proj), "Build should have succeeded.");
-
-				using (var assembly = AssemblyDefinition.ReadAssembly (Path.Combine (Root, b.ProjectDirectory, proj.IntermediateOutputPath, assemblyPath))) {
-					Assert.IsTrue (assembly.MainModule.GetType (handlerType) != null, $"'{handlerTypeFullName}' should have been preserved by the linker.");
-				}
-			}
-		}
-
-		[Test]
-		public void PreserveCustomHttpClientHandlers ([Values (TrimMode.Partial, TrimMode.Full)] TrimMode trimMode, [Values] AndroidRuntime runtime)
-		{
-			PreserveCustomHttpClientHandler ("Xamarin.Android.Net.AndroidMessageHandler", "",
-				$"temp/PreserveAndroidMessageHandler{trimMode}{runtime}", "android-arm64/linked/Mono.Android.dll", trimMode, runtime);
-			PreserveCustomHttpClientHandler ("System.Net.Http.SocketsHttpHandler", "System.Net.Http",
-				$"temp/PreserveSocketsHttpHandler{trimMode}{runtime}", "android-arm64/linked/System.Net.Http.dll", trimMode, runtime);
-			PreserveCustomHttpClientHandler ("MyCustomHandler", "MyClassLibrary",
-				$"temp/MyCustomHandler{trimMode}{runtime}", "android-arm64/linked/MyClassLibrary.dll", trimMode, runtime);
-		}
-
 		[Test]
 		public void WarnAboutAppDomains ([Values] bool isRelease, [Values] AndroidRuntime runtime)
 		{
